Mods to run on python 3.5
in v3.5, plots are not popout, and implicit scope of numpy and matplot functions did not work  (e.g. cos(theta) didn't work without the "np." prefix.  Worse, some other items (sum) are overloaded, and "work" but do something different and undesired in the non numpy version. This was insidious.  grr.